requirements.txt
APScheduler==3.7.0
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
